% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm_repository}
\alias{dgm_repository}
\title{Return OSF Link to the DGM Repository}
\usage{
dgm_repository(dgm_name)
}
\arguments{
\item{dgm_name}{Character string specifying the DGM type}
}
\value{
A character containing an OSF link.
}
\description{
This function returns a string with the link to the OSF project containing
simulated data sets and methods' results.
}
\examples{
dgm_conditions("no_bias")

}
