% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm_settings}
\alias{dgm_settings}
\title{Return Pre-specified DGM Settings}
\usage{
dgm_settings(dgm_name)
}
\arguments{
\item{dgm_name}{Character string specifying the DGM type}
}
\value{
A data frame containing the pre-specified settings including a
\code{setting_id} column which maps settings id to the corresponding settings.
}
\description{
This function returns the list of pre-specified settings for a given Data
Generating Mechanism (DGM).
}
\examples{
dgm_settings("Carter2019")

dgm_settings("Alinaghi2018")

dgm_settings("Stanley2017")

}
