% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm-Carter2019.R
\name{.HongAndReed2021_Carter2019_simMA}
\alias{.HongAndReed2021_Carter2019_simMA}
\title{param k the number of studies in the MA
param delta the true effect (or the average of the true effects if heterogeneity exists)
param tau the SD around the true effect
param censor The censoring function - either "none", "med" (medium publication bias), "high" (high publication bias), or a vector of 3 values for the censoring function (posSign_NS_baseRate, negSign_NS_baseRate, counterSig_rate)
param qrpEnv the qrp environment that produced the literature: 'none', 'low', 'med', 'high'
param empN.boost A constant that is added to the empirical effect sizes: WARNING: NOT CAREFULLY TESTED YET!!}
\usage{
.HongAndReed2021_Carter2019_simMA(
  k,
  delta,
  tau,
  qrpEnv = c("none", "low", "medium", "high"),
  censorFunc = c("none", "medium", "high"),
  verbose = FALSE,
  fixed.n = NULL
)
}
\description{
param k the number of studies in the MA
param delta the true effect (or the average of the true effects if heterogeneity exists)
param tau the SD around the true effect
param censor The censoring function - either "none", "med" (medium publication bias), "high" (high publication bias), or a vector of 3 values for the censoring function (posSign_NS_baseRate, negSign_NS_baseRate, counterSig_rate)
param qrpEnv the qrp environment that produced the literature: 'none', 'low', 'med', 'high'
param empN.boost A constant that is added to the empirical effect sizes: WARNING: NOT CAREFULLY TESTED YET!!
}
