% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measure_mcse}
\alias{measure_mcse}
\title{Get performance measure MCSE function}
\usage{
measure_mcse(measure, ...)
}
\arguments{
\item{measure}{Character string specifying the measure name. If missing, returns a vector of all available measures.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A function that computes the MCSE of the requested measure, or a character vector of measure names.
}
\description{
Returns the function for computing the Monte Carlo Standard Error (MCSE) of a specific performance measure.
Can also be used to list available measures with MCSE functions.
}
