% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method.default}
\alias{method.default}
\title{Default method handler}
\usage{
\method{method}{default}(method_name, data, settings = list())
}
\arguments{
\item{method_name}{Character string specifying the method type}

\item{data}{Data frame containing yi (effect sizes) and sei (standard errors)}

\item{settings}{Either a character identifying a method version or list
containing method-specific settings. An emty input will result in running the
default (first implemented) version of the method.}
}
\value{
Throws an error indicating the method type is unknown. This default
method is only called when no specific method implementation is found for the
given \code{method_name}.
}
\description{
Default method handler
}
