% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mean.R
\name{method.mean}
\alias{method.mean}
\title{Mean Method}
\usage{
\method{method}{mean}(method_name, data, settings)
}
\arguments{
\item{method_name}{Method name (automatically passed)}

\item{data}{Data frame with yi (effect sizes)}

\item{settings}{List of method settings (see Details)}
}
\value{
Data frame with mean results
}
\description{
Implements the unweighted mean method. I.e., the mean of observed effect sizes.
}
\details{
The following settings are implemented \describe{
\item{\code{"default"}}{No settings}
}
}
\examples{
# Generate some example data
data <- data.frame(
  yi = c(0.2, 0.3, 0.1, 0.4, 0.25),
  sei = c(0.1, 0.15, 0.08, 0.12, 0.09)
)

# Apply mean method
result <- run_method("mean", data)
print(result)

}
\references{
\insertAllCited{}
}
\author{
František Bartoš \email{f.bartos96@gmail.com}
}
