% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method_extra_columns.default}
\alias{method_extra_columns.default}
\title{Default method_extra_columns handler}
\usage{
method_extra_columns.default(method_name)
}
\arguments{
\item{method_name}{Character string of the method name}
}
\description{
Default method_extra_columns handler
}
