% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method_settings}
\alias{method_settings}
\alias{get_method_setting}
\title{Return Pre-specified Method Settings}
\usage{
method_settings(method_name)

get_method_setting(method_name, version_id)
}
\arguments{
\item{method_name}{Character string specifying the method type}

\item{version_id}{which method version should be used.}
}
\value{
A list containing the pre-specified settings. For most methods, the
list contains extension of the function call, however, a more elaborate list
of settings that is dispatched within the method call is possible.
}
\description{
This function returns the list of pre-specified settings for a given Method
}
\examples{
method_settings("RMA")
get_method_setting("RMA", version_id = "default")

}
