% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_dgm_metrics}
\alias{retrieve_dgm_metrics}
\title{Retrieve Pre-Computed Performance Metrics for a DGM}
\usage{
retrieve_dgm_metrics(
  dgm_name,
  metric = NULL,
  method = NULL,
  condition_id = NULL,
  path = getwd()
)
}
\arguments{
\item{dgm_name}{Character string specifying the DGM type}

\item{metric}{Which performance metric should be returned (e.g., "bias", "mse", "coverage").
All metrics can be returned by setting to \code{NULL}.}

\item{method}{Which method should be returned. All methods can be returned by setting to \code{NULL}.}

\item{condition_id}{which conditions should settings be returned for.}

\item{path}{Character string specifying the directory path where the dataset should
be saved. Defaults to the current working directory.}
}
\value{
A data.frame
}
\description{
This function returns pre-computed performance metrics for a specified
Data Generating Mechanism (DGM). The pre-computed metrics must be already stored
locally. See \code{\link[=download_dgm_metrics]{download_dgm_metrics()}} function for more guidance.
}
\examples{
\dontrun{
  # get bias metrics for all methods and conditions
  retrieve_dgm_metrics("no_bias", metric = "bias")

  # get all metrics for RMA method
  retrieve_dgm_metrics("no_bias", method = "RMA")

  # get MSE metrics for PET method in condition 1
  retrieve_dgm_metrics("no_bias", metric = "mse", method = "PET", condition_id = 1)
}

}
