% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{S_G_squared}
\alias{S_G_squared}
\title{Calculate sample variance of generic statistic}
\usage{
S_G_squared(G)
}
\arguments{
\item{G}{Vector of generic statistics}
}
\value{
Sample variance S_G^2
}
\description{
Calculate sample variance of generic statistic
}
\keyword{internal}
