% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{S_theta_minus_theta_squared}
\alias{S_theta_minus_theta_squared}
\title{Calculate sample variance of squared errors}
\usage{
S_theta_minus_theta_squared(theta_hat, theta)
}
\arguments{
\item{theta_hat}{Vector of estimates}

\item{theta}{True parameter value}
}
\value{
Sample variance S_(theta_hat - theta)^2
}
\description{
Calculate sample variance of squared errors
}
\keyword{internal}
