% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{S_theta_squared}
\alias{S_theta_squared}
\title{Calculate sample variance of estimates}
\usage{
S_theta_squared(theta_hat)
}
\arguments{
\item{theta_hat}{Vector of estimates}
}
\value{
Sample variance S_theta^2
}
\description{
Calculate sample variance of estimates
}
\keyword{internal}
