% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{S_w_squared}
\alias{S_w_squared}
\title{Calculate sample variance of CI widths}
\usage{
S_w_squared(ci_upper, ci_lower)
}
\arguments{
\item{ci_upper}{Vector of upper CI bounds}

\item{ci_lower}{Vector of lower CI bounds}
}
\value{
Sample variance S_w^2
}
\description{
Calculate sample variance of CI widths
}
\keyword{internal}
