% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm-Alinaghi2019.R
\name{dgm.Alinaghi2018}
\alias{dgm.Alinaghi2018}
\title{Alinaghi and Reed 2018 Data Generating Model}
\usage{
\method{dgm}{Alinaghi2018}(dgm_name, settings)
}
\arguments{
\item{dgm_name}{DGM name (automatically passed)}

\item{settings}{List containing \describe{
\item{environment}{Type of the simulation environment. One of \code{"FE"},
\code{"RE"}, or \code{"PRE"}.}
\item{mean_effect}{Mean effect}
}}
}
\value{
Data frame with \describe{
\item{yi}{effect size}
\item{sei}{standard error}
\item{ni}{sample size}
\item{study_id}{study identifier}
\item{es_id}{effect size identifier}
}
}
\description{
This data generating model simulates univariate regression studies where a variable X
affects a continuous outcome Y. Each study estimates the coefficient of X, which consists
of a fixed component (α1) representing the overall mean effect, and a random component
that varies across studies but is constant within each study. In the "Random Effects"
environment (\code{"RE"}), each study produces one estimate, and the population effect
differs across studies.

The description and code is based on
\insertCite{hong2021using;textual}{PublicationBiasBenchmark}.
The data generating model was introduced in
\insertCite{alinaghi2018meta;textual}{PublicationBiasBenchmark}.
}
\details{
This data generating model is based on Alinaghi & Reed (2018), who study univariate
regression models where a variable X affects a continuous variable Y. The parameter
of interest is the coefficient on X. In the "Random Effects" environment, each study
produces one estimate, and the population effect differs across studies. The coefficient
on X equals a fixed component (α1) plus a random component that is fixed within a study
but varies across studies. The overall mean effect of X on Y is given by α1.

A distinctive feature of Alinaghi & Reed's experiments is that the sample size of
estimated effects is fixed before publication selection, making the meta-analyst's
sample size endogenous and affected by the effect size. Large population effects
are subject to less publication selection, as most estimates satisfy the selection
criteria (statistical significance or correct sign).

Another feature is the separation of statistical significance and sign of the estimated
effect as criteria for selection. Significant/correctly-signed estimates are always
"published," while insignificant/wrong-signed estimates have only a 10\% chance of
being published. This allows for different and sometimes conflicting consequences for
estimator performance.

The simulations are designed to be representative of meta-analyses in economics and
business, which typically have several hundred estimates and substantial effect
heterogeneity. In addition to the "Random Effects" environment, a
"Panel Random Effects" environment is included, where each study has 10 estimates,
modeling the common scenario of multiple estimates per study. Effect estimates and
standard errors are simulated to be more similar within studies than across studies,
and publication selection targets the study rather than individual estimates. For
inclusion in the meta-analyst's sample, a study must have at least 7 out of 10 estimates
that are significant or correctly signed.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=dgm]{dgm()}}, \code{\link[=validate_dgm_setting]{validate_dgm_setting()}}
}
