% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm.default}
\alias{dgm.default}
\title{Default DGM handler}
\usage{
\method{dgm}{default}(dgm_name, settings)
}
\arguments{
\item{dgm_name}{Character string specifying the DGM type}

\item{settings}{List containing the required parameters for the DGM or
numeric setting_id}
}
\description{
Default DGM handler
}
