% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm-no_bias.R
\name{dgm.no_bias}
\alias{dgm.no_bias}
\title{Normal DGM (No Publication Bias)}
\usage{
\method{dgm}{no_bias}(dgm_name, settings)
}
\arguments{
\item{dgm_name}{DGM name (automatically passed)}

\item{settings}{List containing n_studies, mean_effect, heterogeneity}
}
\value{
Data frame with generated data
}
\description{
Generates effect sizes from a normal distribution without publication bias
}
