% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm}
\alias{dgm}
\title{Generic data generating mechanism function}
\usage{
dgm(dgm_name, settings)
}
\arguments{
\item{dgm_name}{Character string specifying the DGM type}

\item{settings}{List containing the required parameters for the DGM or numeric settings_id}
}
\value{
A data frame containing the generated data with standardized structure
}
\description{
This function provides a unified interface to various data generating
mechanisms for simulation studies. The specific DGM is determined by
the first argument.
}
\seealso{
\code{\link[=validate_dgm_settigns]{validate_dgm_settigns()}},
\code{\link[=dgm.Stanley2017]{dgm.Stanley2017()}},
\code{\link[=dgm.Alinaghi2019]{dgm.Alinaghi2019()}},
\code{\link[=dgm.Bom2019]{dgm.Bom2019()}}
\code{\link[=Carter2019]{Carter2019()}}
}
