% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_dgm}
\alias{download_dgm}
\alias{download_dgm_datasets}
\alias{download_dgm_results}
\title{Download Datasets of a DGM}
\usage{
download_dgm_datasets(dgm_name, path = getwd(), add = FALSE)

download_dgm_results(dgm_name, path = getwd(), add = FALSE)
}
\arguments{
\item{dgm_name}{Character string specifying the name of the DGM dataset to download.}

\item{path}{Character string specifying the directory path where the dataset should
be saved. Defaults to the current working directory.}

\item{add}{Logical indicating whether to add the downloaded data to an existing dataset.
Defaults to \code{FALSE}.}
}
\value{
\code{TRUE} if the download was successful, otherwise an error is raised.
}
\description{
This function downloads datasets of a specified Data Generating Mechanism (DGM).
}
\examples{
\dontrun{
  download_dgm_datasets("no_bias")
}

}
