% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\title{Method Method}
\usage{
method(method_name, data, settings)
}
\arguments{
\item{method_name}{Character string specifying the method type}

\item{data}{Data frame containing yi (effect sizes) and sei (standard errors)}

\item{settings}{Either a character identifying a method version or list
containing method-specific settings. An emty input will result in running the
default (first implemented) version of the method.}
}
\description{
Method Method
}
