% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\docType{data}
\name{method_extra_columns.default}
\alias{method_extra_columns.default}
\title{Default method_extra_columns handler}
\format{
An object of class \code{character} of length 0.
}
\usage{
method_extra_columns.default
}
\description{
Default method_extra_columns handler
}
\keyword{datasets}
