% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{package}
\name{PublicationBiasBenchmark-package}
\alias{PublicationBiasBenchmark}
\alias{PublicationBiasBenchmark-package}
\title{PublicationBiasBenchmark: Benchmark for Publication Bias Correction Methods}
\description{
Implements a unified interface for benchmarking meta-analytic publication bias correction methods through simulation studies. It provides 1) predefined data-generating mechanisms from the literature, 2) functions for running meta-analytic methods on simulated data, 3) pre-simulated datasets and pre-computed results for reproducible benchmarks, 4) tools for visualizing and comparing method performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FBartos/PublicationBiasBenchmark}
  \item \url{https://fbartos.github.io/PublicationBiasBenchmark/}
  \item Report bugs at \url{https://github.com/FBartos/PublicationBiasBenchmark/issues}
}

}
\author{
\strong{Maintainer}: Franti<U+0161>ek Barto<U+0161> \email{f.bartos96@gmail.com} (\href{https://orcid.org/0000-0002-0018-5573}{ORCID})

Authors:
\itemize{
  \item Samuel Pawel (\href{https://orcid.org/0000-0003-2779-320X}{ORCID})
  \item Bj<U+00F6>rn S. Siepe (\href{https://orcid.org/0000-0002-9558-4648}{ORCID})
}

}
\keyword{internal}
