% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PublicationBiasBenchmark_options}
\alias{PublicationBiasBenchmark_options}
\alias{PublicationBiasBenchmark.options}
\alias{PublicationBiasBenchmark.get_option}
\title{Options for the PublicationBiasBenchmark package}
\usage{
PublicationBiasBenchmark.options(...)

PublicationBiasBenchmark.get_option(name)
}
\arguments{
\item{...}{named option(s) to change - for a list of available options, see
details below.}

\item{name}{the name of the option to get the current value of - for a list of
available options, see details below.}
}
\value{
The current value of all available PublicationBiasBenchmark options (after applying any
changes specified) is returned invisibly as a named list.
}
\description{
A placeholder object and functions for the PublicationBiasBenchmark package.
}
\details{
\describe{
\item{\code{"resources_directory"}}{Location where the benchmark data/results/measures are stored}
\item{\code{"prompt_for_download"}}{Whether each file download should ask for explicit approval}
}
}
