% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_dgm_measures}
\alias{retrieve_dgm_measures}
\title{Retrieve Pre-Computed Performance measures for a DGM}
\usage{
retrieve_dgm_measures(
  dgm_name,
  metric = NULL,
  method = NULL,
  condition_id = NULL,
  path = NULL
)
}
\arguments{
\item{dgm_name}{Character string specifying the DGM type}

\item{metric}{Which performance metric should be returned (e.g., "bias", "mse", "coverage").
All measures can be returned by setting to \code{NULL}.}

\item{method}{Which method should be returned. All methods can be returned by setting to \code{NULL}.}

\item{condition_id}{which conditions should settings be returned for.}

\item{path}{Character string specifying the directory path where the datasets/results/measures
should be saved. Defaults to the location specified via
\code{PublicationBiasBenchmark.get_option("simulation_directory")}. The objects are stored
in dgm_name/datasets, dgm_name/results, dgm_name/measures subfolders.}
}
\value{
A data.frame
}
\description{
This function returns pre-computed performance measures for a specified
Data Generating Mechanism (DGM). The pre-computed measures must be already stored
locally. See \code{\link[=download_dgm_measures]{download_dgm_measures()}} function for more guidance.
}
\examples{
\dontrun{
  # get bias measures for all methods and conditions
  retrieve_dgm_measures("no_bias", metric = "bias")

  # get all measures for RMA method
  retrieve_dgm_measures("no_bias", method = "RMA")

  # get MSE measures for PET method in condition 1
  retrieve_dgm_measures("no_bias", metric = "mse", method = "PET", condition_id = 1)
}

}
