% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_dgm_results}
\alias{retrieve_dgm_results}
\title{Retrieve a Pre-Simulated Condition and Repetition From a DGM}
\usage{
retrieve_dgm_results(
  dgm_name,
  method = NULL,
  condition_id = NULL,
  repetition_id = NULL,
  path = getwd()
)
}
\arguments{
\item{dgm_name}{Character string specifying the DGM type}

\item{method}{Which method should be returned. The complete
results can be returned by setting to \code{NULL}.}

\item{condition_id}{which conditions should settings be returned for.}

\item{repetition_id}{Which repetition should be returned. The complete
condition can be returned by setting to either \code{NULL}.}

\item{path}{Character string specifying the directory path where the dataset should
be saved. Defaults to the current working directory.}
}
\value{
A data.frame
}
\description{
This function returns a pre-simulated dataset of a given repetition and
condition from a dgm. The pre-simulated datasets must be already stored
locally. See \code{\link[=download_dgm_datasets]{download_dgm_datasets()}} function for more guidance.
}
\examples{
\dontrun{
  # get condition 1, repetition 1
  retrieve_dgm_results("no_bias", condition_id = 1, repetition_id = 1)

  # get condition 1, all repetitions
  retrieve_dgm_results("no_bias", condition_id = 1)
}


}
